+++ sdmx-dl-cli/src/main/java/sdmxdl/cli/CheckCommand.java
@@ -28,7 +28,7 @@ import java.util.concurrent.Callable;
         subcommands = {
                 CheckStatusCommand.class,
                 CheckAccessCommand.class,
-                CheckPropertiesCommand.class
+                CheckConfigCommand.class
         }
 )
 public final class CheckCommand implements Callable<Void> {