+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Renamed command check.properties as check.config [#136](https://github.com/nbbrd/sdmx-dl/issues/136)
+
 ## [3.0.0-beta.3] - 2021-09-13
 
 This is the third beta release of **sdmx-dl**.